jarvis-github-ready/requirements.txt
fastmcp>=0.2.0
mcp>=1.0.0
google-generativeai>=0.8.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
httpx>=0.27.0
google-api-python-client>=2.100.0
google-auth>=2.25.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
pydantic>=2.0.0
